move serde support into a v1 module
So that we've got some room to support a hypothetical future version
